# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Geo Portal API
  version: 0.1.0
  description: Internal API for the Geo Portal SPA. Read-only over the knowledge base.
paths:
...
  /api/map/assets:
    get:
      operationId: map_assets_at_point
      description: |-
        List the catalog assets covering a clicked point or overlapping a drawn area.

        One endpoint for both because they are one question — "what data is about
        this place" — asked with two geometries, and the answer has the same shape
        either way (docs/adr/014). The operation keeps its original id so the
        generated client's hook keeps its name.
      summary: Assets covering a point or a drawn area
      parameters:
      - in: query
        name: area
        schema:
          type: string
          minLength: 1
        description: A drawn polygon as WKT, POLYGON((lon lat, ...)), SRID 4326, at
          most 100 corners and not crossing itself. Sent instead of lon and lat.
      - in: query
        name: lat
        schema:
          type: number
          format: double
          maximum: 90
          minimum: -90
        description: Latitude, WGS 84. Sent with `lon`, never with `area`.
      - in: query
        name: lon
        schema:
          type: number
          format: double
          maximum: 180
          minimum: -180
        description: Longitude, WGS 84. Sent with `lat`, never with `area`.
      tags:
      - map
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetsAtPoint'
          description: ''
        '400':
          description: No place given, both given, or a point or area that cannot
            be read.
  /api/map/assets/{asset_id}:
    get:
      operationId: map_asset_detail
      description: One asset's metadata and footprint.
      summary: Asset metadata and footprint
      parameters:
      - in: path
        name: asset_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - map
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDetail'
          description: ''
        '404':
          description: No active asset with that id.
  /api/map/assets/{asset_id}/data:
    get:
      operationId: map_asset_data
      description: |-
        One vector asset's features, read from the lake.

        Vector only, and the endpoint says so rather than inferring it: `assets`
        holds every modality and the lake read only makes sense for GeoParquet
        (docs/adr/008). Other modalities get their own endpoints.
      summary: Vector asset features
      parameters:
      - in: query
        name: area
        schema:
          type: string
          minLength: 1
        description: 'Read only the features that intersect this drawn polygon: WKT,
          POLYGON((lon lat, ...)), SRID 4326, at most 100 corners and not crossing
          itself. Matching is exact. Sent instead of bbox.'
      - in: path
        name: asset_id
        schema:
          type: string
          format: uuid
        required: true
      - in: query
        name: bbox
        schema:
          type: string
          minLength: 1
        description: 'Read only the features whose bounding box overlaps this area:
          min_lon,min_lat,max_lon,max_lat, SRID 4326. Matching is on bounding boxes,
          so a feature just outside the area may be included. Without it the whole
          file is in scope.'
      - in: query
        name: cursor
        schema:
          type: integer
          minimum: 0
        description: The `next_cursor` of the previous page.
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
        description: Features per page. Defaults to, and is capped at, the server's
          page size.
      tags:
      - map
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetData'
          description: ''
        '400':
          description: An unreadable bbox, area, cursor or limit, or both bbox and
            area.
        '404':
          description: No active vector asset with that id.
        '502':
          description: The file could not be read.
        '503':
          description: The lake could not be reached.
...
components:
  schemas:
...
    AssetData:
      type: object
      description: |-
        A vector asset's features, read from the lake.

        There is no serializer for a feature: its properties are whatever columns
        the GeoParquet file holds, and they differ per file. The shape is declared
        to the contract instead, by `FeatureListField`, so the generated client
        still gets a typed array rather than `unknown[]`.
      properties:
        asset_id:
          type: string
          format: uuid
          readOnly: true
        count:
          type: integer
          readOnly: true
          description: How many features are in this response.
        next_cursor:
          type: integer
          readOnly: true
          nullable: true
          description: Pass back as `cursor` for the next page of this area. Null
            when this page is the last.
        features:
          type: array
          items:
            type: object
            title: Feature
            properties:
              type:
                type: string
                enum:
                - Feature
              geometry:
                type: object
                nullable: true
                additionalProperties: true
                description: A GeoJSON geometry, SRID 4326.
              properties:
                type: object
                additionalProperties: true
                description: The file's non-geometry columns.
            required:
            - type
            - geometry
            - properties
          readOnly: true
          description: GeoJSON Features, SRID 4326.
      required:
      - asset_id
      - count
      - features
      - next_cursor
    AssetDetail:
      type: object
      description: One asset's dynamic metadata and its footprint.
      properties:
        asset_id:
          type: string
          format: uuid
          readOnly: true
        data_type:
          type: string
          readOnly: true
        metadata:
          type: object
          additionalProperties: {}
          readOnly: true
          description: Whatever the catalog holds for this asset. The key set is not
            fixed and differs by data type.
        footprint:
          readOnly: true
          description: The asset's coverage as a GeoJSON geometry, SRID 4326.
      required:
      - asset_id
      - data_type
      - footprint
      - metadata
    AssetGroup:
      type: object
      description: |-
        Assets sharing a data type.

        The data type is whatever the knowledge base classifies the asset as; the
        portal neither defines nor translates the list (specs/map.md).
      properties:
        data_type:
          type: string
          readOnly: true
        assets:
          type: array
          items:
            $ref: '#/components/schemas/AssetSummary'
          readOnly: true
      required:
      - assets
      - data_type
...
    AssetSummary:
      type: object
      description: One asset as it appears in a point-click list.
      properties:
        asset_id:
          type: string
          format: uuid
          readOnly: true
        source_uri:
          type: string
          readOnly: true
        format:
          type: string
          readOnly: true
          nullable: true
        summary:
          type: string
          readOnly: true
          nullable: true
        time_start:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        time_end:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - asset_id
      - format
      - source_uri
      - summary
      - time_end
      - time_start
    AssetsAtPoint:
      type: object
      description: The whole answer to a point click. No coverage means no groups.
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/AssetGroup'
          readOnly: true
      required:
      - groups
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT